no comment

+++ Cursa/Controllers/ModuleTypesController.cs
@@ -1,30 +1,80 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Linq.Dynamic.Core;
 using System.Threading.Tasks;
+using AutoMapper;
+using Cursa.ViewModels.ModuleTypesVM;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using DataLayer;
 using DataLayer.Entities;
+using Microsoft.Extensions.Logging;
 
 namespace Cursa.Controllers
 {
     public class ModuleTypesController : Controller
     {
         private readonly EfDbContext _context;
+        private readonly IMapper _mapper;
+        private readonly ILogger<ModuleTypesController> _logger;
 
-        public ModuleTypesController(EfDbContext context)
+        public ModuleTypesController(EfDbContext context, IMapper mapper, ILogger<ModuleTypesController> logger)
         {
             _context = context;
+            _mapper = mapper;
+            _logger = logger;
         }
 
         // GET: ModuleTypes
-        public async Task<IActionResult> Index()
+        public IActionResult Index() => View(new ModuleTypesDisplayViewModel());
+
+        [HttpPost]
+        public IActionResult GetModuleTypes()
         {
-            var efDbContext = _context.ModulesTypes.Include(m => m.ModuleSubTypes);
-            return View(await efDbContext.ToListAsync());
+            try
+            {
+                var draw = Request.Form["draw"].FirstOrDefault();
+                var start = Request.Form["start"].FirstOrDefault();
+                var length = Request.Form["length"].FirstOrDefault();
+                var sortColumn = Request
+                    .Form["columns[" + Request.Form["order[0][column]"].FirstOrDefault() + "][name]"].FirstOrDefault();
+                var sortColumnDirection = Request.Form["order[0][dir]"].FirstOrDefault();
+                var searchGlobalValue = Request.Form["search[value]"].FirstOrDefault();
+                var pageSize = length != null ? Convert.ToInt32(length) : 0;
+                var skip = start != null ? Convert.ToInt32(start) : 0;
+
+                var projectsData = _mapper.ProjectTo<ModuleTypesDisplayViewModel>(_context.ModulesTypes
+                    .AsNoTracking());
+
+                if (!(string.IsNullOrEmpty(sortColumn) && string.IsNullOrEmpty(sortColumnDirection)))
+                {
+                    projectsData = projectsData.OrderBy(sortColumn + " " + sortColumnDirection);
+                }
+
+                if (!string.IsNullOrEmpty(searchGlobalValue))
+                {
+                    projectsData = projectsData.Where(m => m.Name.Contains(searchGlobalValue)
+                                                           || m.Code.Contains(searchGlobalValue)
+                                                           || m.NumberConnectionPoints.ToString()
+                                                               .Contains(searchGlobalValue)
+                                                           || m.CountChanel.ToString().Contains(searchGlobalValue));
+                }
+
+                var recordsTotal = projectsData.Count();
+                var data = projectsData.Skip(skip).Take(pageSize).ToList();
+                var jsonData = new
+                    {draw = draw, recordsFiltered = recordsTotal, recordsTotal = recordsTotal, data = data};
+                return Ok(jsonData);
+            }
+            catch (Exception e)
+            {
+                _logger.LogError("Error search:{ExceptionMessage}", e.Message);
+                return NotFound();
+            }
         }
+
 
         // GET: ModuleTypes/Details/5
         public async Task<IActionResult> Details(int? id)
@@ -35,7 +85,6 @@ namespace Cursa.Controllers
             }
 
             var moduleType = await _context.ModulesTypes
-                .Include(m => m.ModuleSubTypes)
                 .FirstOrDefaultAsync(m => m.Id == id);
             if (moduleType == null)
             {
@@ -48,7 +97,6 @@ namespace Cursa.Controllers
         // GET: ModuleTypes/Create
         public IActionResult Create()
         {
-            ViewData["ModuleSubTypesId"] = new SelectList(_context.ModulesSubTypes, "Id", "Name");
             return View();
         }
 
@@ -57,7 +105,10 @@ namespace Cursa.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Name,Code,CountChanel,ModuleSubTypesId,Description,CreatedDate,Id")] ModuleType moduleType)
+        public async Task<IActionResult> Create(
+            [Bind(
+                "Name,Code,CountChanel,IsActiv,IsCommunicationDevice,NumberConnectionPoints,Description,CreatedDate,Id")]
+            ModuleType moduleType)
         {
             if (ModelState.IsValid)
             {
@@ -65,7 +116,7 @@ namespace Cursa.Controllers
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["ModuleSubTypesId"] = new SelectList(_context.ModulesSubTypes, "Id", "Name", moduleType.ModuleSubTypesId);
+
             return View(moduleType);
         }
 
@@ -82,7 +133,7 @@ namespace Cursa.Controllers
             {
                 return NotFound();
             }
-            ViewData["ModuleSubTypesId"] = new SelectList(_context.ModulesSubTypes, "Id", "Name", moduleType.ModuleSubTypesId);
+
             return View(moduleType);
         }
 
@@ -91,7 +142,10 @@ namespace Cursa.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Name,Code,CountChanel,ModuleSubTypesId,Description,CreatedDate,Id")] ModuleType moduleType)
+        public async Task<IActionResult> Edit(int id,
+            [Bind(
+                "Name,Code,CountChanel,IsActiv,IsCommunicationDevice,NumberConnectionPoints,Description,CreatedDate,Id")]
+            ModuleType moduleType)
         {
             if (id != moduleType.Id)
             {
@@ -116,9 +170,10 @@ namespace Cursa.Controllers
                         throw;
                     }
                 }
+
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["ModuleSubTypesId"] = new SelectList(_context.ModulesSubTypes, "Id", "Name", moduleType.ModuleSubTypesId);
+
             return View(moduleType);
         }
 
@@ -131,7 +186,6 @@ namespace Cursa.Controllers
             }
 
             var moduleType = await _context.ModulesTypes
-                .Include(m => m.ModuleSubTypes)
                 .FirstOrDefaultAsync(m => m.Id == id);
             if (moduleType == null)
             {
@@ -147,9 +201,24 @@ namespace Cursa.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var moduleType = await _context.ModulesTypes.FindAsync(id);
-            _context.ModulesTypes.Remove(moduleType);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            if (moduleType == null)
+            {
+                return NotFound();
+            }
+
+            try
+            {
+                _context.ModulesTypes.Remove(moduleType);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            catch (DbUpdateException e)
+            {
+                _logger.LogInformation("{ExceptionMessage}", e.Message);
+                ModelState.AddModelError(String.Empty, "Невозможно удалить, тип модуля используется в системе");
+            }
+
+            return View(moduleType);
         }
 
         private bool ModuleTypeExists(int id)
